working on libs as plugin
