+++ src/main/groovy/org/boozallen/plugins/jte/config/LibraryProviderDescriptor.groovy
@@ -1,0 +1,23 @@
+package org.boozallen.plugins.jte.config
+
+import org.boozallen.plugins.jte.utils.FileSystemWrapper
+import org.boozallen.plugins.jte.console.TemplateLogger
+import org.boozallen.plugins.jte.binding.injectors.LibraryLoader
+import org.kohsuke.stapler.DataBoundConstructor
+import org.kohsuke.stapler.DataBoundSetter
+import hudson.scm.SCM
+import jenkins.scm.api.SCMFileSystem
+import jenkins.scm.api.SCMFile 
+import hudson.Extension
+import hudson.model.AbstractDescribableImpl
+import hudson.model.Descriptor
+import hudson.Util
+import org.jenkinsci.plugins.workflow.cps.CpsScript
+import hudson.ExtensionPoint
+import hudson.ExtensionList 
+import jenkins.model.Jenkins 
+
+
+public abstract class LibraryProviderDescriptor extends Descriptor<LibraryProvider> {
+    String getDisplayName(){}
+}
