+++ src/main/resources/org/boozallen/plugins/jte/config/GovernanceTier/config.jelly
@@ -23,13 +23,9 @@
     <f:entry field="baseDir" title="${%Configuration Base Directory}">
         <f:textbox default=""/>
     </f:entry>
-
-    <f:block>
-        <j:out value="${%Library Sources}"/>
-    </f:block>
     <f:entry>
         <f:repeatableProperty field="librarySources" header="${%Library}">
-            <f:entry title="">
+            <f:entry title="Library Source">
                 <div align="right">
                     <f:repeatableDeleteButton/>
                 </div>
